+++ flash/core/data/data_source.py
@@ -163,6 +163,7 @@ class DefaultDataSources(LightningEnum):
     DATAFRAME = "data_frame"
     LISTS = "lists"
     SENTENCES = "sentences"
+    LABELSTUDIO = "labelstudio"
 
     # TODO: Create a FlashEnum class???
     def __hash__(self) -> int: